moved sui sdk implementation from index file
