+++ sui/src/sui.ts
@@ -1,0 +1,231 @@
+/**
+ * Copyright (c) 2025, Everstake.
+ * Licensed under the BSD-3-Clause License. See LICENSE file for details.
+ */
+
+import { DelegatedStake, SuiClient } from '@mysten/sui/client';
+import { Blockchain } from './utils';
+import {
+  SUI_BASE_NUM,
+  SUI_MIN_AMOUNT_FOR_STAKE,
+  SUI_NETWORK_ADDRESSES,
+} from './constants';
+import { ERROR_MESSAGES, ORIGINAL_ERROR_MESSAGES } from './constants/error';
+import { SuiNetworkType } from './types';
+import BigNumber from 'bignumber.js';
+import {
+  isValidSuiAddress,
+  SUI_SYSTEM_STATE_OBJECT_ID,
+} from '@mysten/sui/utils';
+import { Transaction } from '@mysten/sui/transactions';
+
+export class Sui extends Blockchain {
+  protected ERROR_MESSAGES = ERROR_MESSAGES;
+  protected ORIGINAL_ERROR_MESSAGES = ORIGINAL_ERROR_MESSAGES;
+
+  public validatorAddress!: string;
+  public client!: SuiClient;
+
+  constructor(network: SuiNetworkType = 'mainnet', url?: string) {
+    super();
+    this.initializeNetwork(network, url);
+  }
+
+  /**
+   * Selects and initializes a new network.
+   *
+   * This method calls `initializeNetwork` with the provided parameters and returns the current instance,
+   * allowing for method chaining.
+   *
+   * @param network - The network type. This should be one of the keys in `SUI_NETWORK_ADDRESSES`.
+   * @param url - The RPC URL of the network. If not provided, the method will use the URL from `SUI_NETWORK_ADDRESSES`.
+   *
+   * @returns The current instance of the `Sui` class.
+   */
+  public selectNetwork(network: SuiNetworkType, url?: string): Sui {
+    this.initializeNetwork(network, url);
+
+    return this;
+  }
+
+  /**
+   * Initializes the network.
+   *
+   * This method sets the validator address and initializes the SuiClient with the appropriate RPC URL.
+   *
+   * @param network - The network type. This should be one of the keys in `SUI_NETWORK_ADDRESSES`.
+   * @param url - The RPC URL of the network. If not provided, the method will use the URL from `SUI_NETWORK_ADDRESSES`.
+   *
+   * @throws Will throw an error if the provided network is not supported (i.e., not a key in `SUI_NETWORK_ADDRESSES`).
+   */
+  private initializeNetwork(network: SuiNetworkType, url?: string) {
+    const networkAddresses = SUI_NETWORK_ADDRESSES[network];
+
+    if (!networkAddresses) {
+      this.throwError('NETWORK_NOT_SUPPORTED', network);
+    }
+
+    this.validatorAddress = networkAddresses.validatorAddress;
+    this.client = new SuiClient({
+      url: url ?? networkAddresses.rpcUrl,
+    });
+  }
+
+  /**
+   * Retrieves the delegated stakes for a given address.
+   *
+   * @param address - The address to fetch staking balances for.
+   *
+   * @returns A promise that resolves to an array of DelegatedStake objects.
+   *
+   * @throws Will throw an error if the address is not valid or if the API call fails.
+   */
+  public async getStakeBalanceByAddress(
+    address: string,
+  ): Promise<DelegatedStake[]> {
+    if (!this.isAddress(address)) {
+      this.throwError('ADDRESS_FORMAT_ERROR');
+    }
+
+    try {
+      const delegatedStakes = await this.client.getStakes({ owner: address });
+
+      return delegatedStakes;
+    } catch (error) {
+      throw this.handleError('STAKE_BALANCE_ERROR', error);
+    }
+  }
+
+  /**
+   * Retrieves the Sui balance for a given address.
+   *
+   * @param address - The address to fetch the balance for.
+   *
+   * @returns A promise that resolves to a BigNumber representing the balance.
+   *
+   * @throws Will throw an error if the address is not valid or if the API call fails.
+   */
+  public async getBalanceByAddress(address: string): Promise<BigNumber> {
+    if (!this.isAddress(address)) {
+      this.throwError('ADDRESS_FORMAT_ERROR');
+    }
+
+    try {
+      const balance = await this.client.getBalance({ owner: address });
+
+      return new BigNumber(balance.totalBalance);
+    } catch (error) {
+      throw this.handleError('USER_BALANCES_ERROR', error);
+    }
+  }
+
+  /**
+   * Creates a transaction to transfer SUI tokens to a recipient.
+   *
+   * @param recipientAddress - The address of the recipient.
+   * @param amount - The amount of SUI tokens to transfer as a string.
+   *
+   * @returns A promise that resolves to a Transaction object ready to be signed and executed.
+   *
+   * @throws Will throw an error if the recipient address is not valid or if the transaction preparation fails.
+   */
+  public async sendTransfer(
+    recipientAddress: string,
+    amount: string,
+  ): Promise<Transaction> {
+    if (!this.isAddress(recipientAddress)) {
+      this.throwError('ADDRESS_FORMAT_ERROR');
+    }
+
+    try {
+      const tx = new Transaction();
+
+      const coin = tx.splitCoins(tx.gas, [
+        SUI_BASE_NUM.multipliedBy(amount).toString(),
+      ]);
+
+      tx.transferObjects([coin], recipientAddress);
+
+      return tx;
+    } catch (error) {
+      throw this.handleError('TRANSFER_ERROR', error);
+    }
+  }
+
+  /**
+   * Creates a transaction to stake SUI tokens with a validator.
+   *
+   * @param amount - The amount of SUI tokens to stake as a string.
+   *
+   * @returns A promise that resolves to a Transaction object ready to be signed and executed.
+   *
+   * @throws Will throw an error if the amount is less than the minimum required for staking
+   * or if the transaction preparation fails.
+   */
+  public async stake(amount: string) {
+    if (+amount < SUI_MIN_AMOUNT_FOR_STAKE) {
+      this.throwError('MIN_STAKE_AMOUNT_ERROR');
+    }
+
+    try {
+      const tx = new Transaction();
+      const stakeCoin = tx.splitCoins(tx.gas, [
+        SUI_BASE_NUM.multipliedBy(amount).toString(),
+      ]);
+
+      tx.moveCall({
+        target: '0x3::sui_system::request_add_stake',
+        arguments: [
+          tx.sharedObjectRef({
+            objectId: SUI_SYSTEM_STATE_OBJECT_ID,
+            initialSharedVersion: 1,
+            mutable: true,
+          }),
+          stakeCoin,
+          tx.pure.address(this.validatorAddress),
+        ],
+      });
+
+      return tx;
+    } catch (error) {
+      throw this.handleError('STAKE_ERROR', error);
+    }
+  }
+
+  /**
+   * Creates a transaction to unstake SUI tokens from a validator.
+   *
+   * @param stakedSuiId - The ID of the staked SUI object to withdraw.
+   *
+   * @returns A promise that resolves to a Transaction object ready to be signed and executed.
+   *
+   * @throws Will throw an error if the transaction preparation fails.
+   */
+  public async unstake(stakedSuiId: string): Promise<Transaction> {
+    try {
+      const tx = new Transaction();
+      tx.moveCall({
+        target: '0x3::sui_system::request_withdraw_stake',
+        arguments: [
+          tx.object(SUI_SYSTEM_STATE_OBJECT_ID),
+          tx.object(stakedSuiId),
+        ],
+      });
+
+      return tx;
+    } catch (error) {
+      throw this.handleError('UNSTAKE_ERROR', error);
+    }
+  }
+
+  /**
+   * Validates if a string is a valid Sui address.
+   *
+   * @param address - The string to validate as a Sui address.
+   *
+   * @returns A boolean indicating whether the string is a valid Sui address.
+   */
+  private isAddress(address: string): boolean {
+    return isValidSuiAddress(address);
+  }
+}
